+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/util/Tactics.java
@@ -1,5 +1,7 @@
 package net.picopress.mc.mods.zombietactics2.util;
 
+import net.picopress.mc.mods.zombietactics2.impl.Plane;
+
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Vec3i;
@@ -23,7 +25,7 @@ import net.minecraft.world.level.block.Rotation;
 import net.minecraft.world.level.chunk.LevelChunk;
 import net.minecraft.world.phys.AABB;
 
-import net.picopress.mc.mods.zombietactics2.impl.Plane;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -44,21 +46,34 @@ public class Tactics {
         else return Rotation.COUNTERCLOCKWISE_90; // x = 1, z = 0
     }
 
-
-    public static float getExactDamage(ServerLevel serverLevel, Mob mob, LivingEntity target) {
+    public static float getExactDamage(@NotNull Mob mob, LivingEntity target) {
+        ServerLevel serverLevel = getSl(mob);
         var tmp = mob.getAttribute(Attributes.ATTACK_DAMAGE);
         if(tmp == null) return 0;
         float dam = (float)tmp.getValue();
+        if(serverLevel == null) return dam;
         return EnchantmentHelper.modifyDamage(serverLevel, mob.getWeaponItem(), target, mob.damageSources().mobAttack(mob), dam);
     }
 
     // thonk
     public static class Heuristic {
+        /**
+         * calculates the power of the mob
+         * @param target the target entity
+         * @return the power of the mob, which is calculated as (attack + 1) * (health / 5) * speed + 1
+         */
         public static int getEnemyPower(LivingEntity target) {
             var attack = target.getAttribute(Attributes.ATTACK_DAMAGE);
             return (int)((attack != null? attack.getValue(): 0) / 2 * target.getHealth() / 5 * target.getSpeed() + 1);
         }
 
+        /**
+         *
+         * @param clazz the class of the mob
+         * @param mob attacker
+         * @param target the target entity
+         * @return true if the mob can win against the target, false otherwise
+         */
         public static boolean simulate(Class<? extends LivingEntity> clazz, Mob mob, LivingEntity target) {
             // friends list
             var peers = mob.level().getEntitiesOfClass(clazz, ((Plane)mob).zombietactics2$getFollowingArea(), (liv) -> liv != mob);
@@ -86,10 +101,23 @@ public class Tactics {
     }
 
     public static class ItemUtil {
+        /**
+         * returns the defense point
+         * @param armor armor item
+         * @return the defense point of the armor, which is calculated as (defense + 1) * (toughness + 1)
+         */
         static double getDefensePoint(ArmorItem armor) {
             return (armor.getDefense() + 1) * (armor.getToughness() + 1);
         }
 
+        /**
+         * it retrieves an attribute modifier of the item stack by its path and namespace.
+         * you should check if it is null
+         * @param stack an item stack
+         * @param path the path of the attribute, e.g. "generic.attack_damage"
+         * @param namespace the namespace of the attribute, e.g. "minecraft"
+         * @return the attribute modifier of the item stack, or null if not found
+         */
         public static @Nullable AttributeModifier getItemAttr(ItemStack stack, String path, String namespace) {
             ItemAttributeModifiers component = stack.getComponents().get(DataComponents.ATTRIBUTE_MODIFIERS);
             if(component == null) return null;
@@ -106,11 +134,22 @@ public class Tactics {
             return entry.modifier();
         }
 
-        // default namespace
+        /**
+         *
+         * @param stack an item stack
+         * @param path the path of the attribute with the namespace "minecraft"
+         * @return the attribute modifier of the item stack, or null if not found
+         */
         static public @Nullable AttributeModifier getItemAttr(ItemStack stack, String path) {
             return getItemAttr(stack, path, "minecraft");
         }
 
+        /**
+         *
+         * @param me the mob holding the item
+         * @param dropped dropped stuff
+         * @return true if the dropped item is better than the mob's one, false otherwise
+         */
         public static boolean isBetter(Mob me, ItemStack dropped) {
             // selecting a weapon
             var my = me.getMainHandItem();
@@ -155,6 +194,16 @@ public class Tactics {
         }
     }
 
+    public static ServerLevel getSl(Mob mob) {
+        var stuff = mob.getServer();
+        return stuff != null? stuff.getLevel(mob.level().dimension()): null; // I'm not in the server
+    }
+
+    // alias
+    public static ServerLevel getServerLevel(Mob mob) {
+        return getSl(mob);
+    }
+
     public static class World {
         // chunk xz = 16*16
         public static LevelChunk[] getNearbyChunks(Level level, BlockPos pos) {
@@ -196,4 +245,6 @@ public class Tactics {
             return Math.abs(p1.getX() - p2.getX()) + Math.abs(p1.getY() - p2.getY()) + Math.abs(p1.getZ() - p2.getZ());
         }
     }
+
+    private Tactics() {}
 }